# textproto: Reduce amount of allocations where possible