Several preallocations and one clever trick.

Here is comparison with old code using benchmarks added in d569c5e:
```
benchmark                            old ns/op     new ns/op     delta
BenchmarkTextprotoReadHeader/1-4     16675         12950         -22.34%
BenchmarkTextprotoReadHeader/2-4     21488         16850         -21.58%
BenchmarkTextprotoReadHeader/3-4     36715         31491         -14.23%

benchmark                            old allocs     new allocs     delta
BenchmarkTextprotoReadHeader/1-4     165            106            -35.76%
BenchmarkTextprotoReadHeader/2-4     183            116            -36.61%
BenchmarkTextprotoReadHeader/3-4     359            277            -22.84%

benchmark                            old bytes     new bytes     delta
BenchmarkTextprotoReadHeader/1-4     16494         14249         -13.61%
BenchmarkTextprotoReadHeader/2-4     26270         21622         -17.69%
BenchmarkTextprotoReadHeader/3-4     24978         30808         +23.34%
```

Header 1 is "the most typical", header 2 contains several very big fields, header 3 contains a lot of small fields.

For reference, here are results using net/textproto parser:
```
BenchmarkStdlibReadHeader/1-4 	  9779 ns/op	    6224 B/op	      38 allocs/op
BenchmarkStdlibReadHeader/2-4 	  12348 ns/op	   12208 B/op	      41 allocs/op
BenchmarkStdlibReadHeader/3-4 	  21608 ns/op	   12304 B/op	     109 allocs/op
```

What I think is worth exploring is possibility of deferring trimAroundNewlines to Header.Get/Fields.Value time. This could avoid extra allocation for fields we are not actually interested in. With header fields littered with server tracing information, this can be useful (side note: headers I used in benchmarks are from actual messages l received).