@foxcpp 2019-10-13
Oh, I think I might to take a look at how net/textproto makes approximation of fields count in header. Since textproto.ReadHeader is also used for multipart headers, it might be a good idea to not preallocate a lot of fields for them.